+++ theme/hub-theme/footer.php
@@ -1,0 +1,12 @@
+<?php
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+?>
+
+<?php hub_theme_region('footer'); ?>
+
+<?php wp_footer(); ?>
+</body>
+</html>